refactor(hooks): promote useResizable to src/hooks
Move useResizable from src/features/workspace/hooks/ to src/hooks/ so
the new global Sidebar can use it without importing from features.

Apply initial-clamp fix: useState lazy initializer ensures out-of-range
defaults are clamped on mount before first drag. Add 2 regression tests
for above-max and below-min initial values.

Update 4 import paths: WorkspaceView.tsx, BottomDrawer.tsx, and
WorkspaceView.command-palette.test.tsx (both vi.mock and dynamic import).
